+++ src/nasem_dairy/__init__.py
@@ -6,7 +6,7 @@ from nasem_dairy.model.utility import read_csv_input, read_json_input, demo, get
 from nasem_dairy.model_output.ModelOutput import ModelOutput
 from nasem_dairy.model.nasem import nasem
 from nasem_dairy.data.constants import coeff_dict, infusion_dict, MP_NP_efficiency_dict, mPrt_coeff_list, f_Imb
-from nasem_dairy.dag.dag import ModelDAG
+from nasem_dairy.dag.ModelDAG import ModelDAG
 from nasem_dairy.nasem_equations.dry_matter_intake import (
     calculate_Kb_LateGest_DMIn,
     calculate_An_PrePartWklim,
